requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
google-genai